+++ app/views/info.njk
@@ -56,7 +56,7 @@
         firstCellIsHeader: false,
         head: [
           {
-            text: "Import task name (data for the last 3 days)"
+            text: "Import task name (data for the last 24 hrs)"
           },
           {
             text: "Completed"
